# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: 'ToggleBox: Remote Config & Feature Flags API'
  description: |
    **ToggleBox** is a multi-platform remote configuration and feature flag management service.

    ## Key Features

    - **Remote Configuration**: Version-controlled, immutable config management
    - **Feature Flags**: Flexible toggles with percentage and targeted rollouts
    - **Multi-Platform**: Separate configs for web, mobile, and other platforms
    - **Environment Management**: Staging, production, and custom environments
    - **CDN Caching**: CloudFront-backed caching with cache invalidation
    - **Client SDKs**: JavaScript, Next.js, and Expo/React Native

    ## API Structure

    ### Public Endpoints (Read-Only)
    - **Path**: `/api/v1/platforms/*`
    - **Access**: Publicly accessible, no authentication
    - **Methods**: GET only
    - **Caching**: CloudFront CDN with configurable TTL

    ### Authentication Endpoints
    - **Paths**: `/api/v1/auth/*`, `/api/v1/users/*`, `/api/v1/api-keys/*`
    - **Access**: Public for registration/login, authenticated for management
    - **Methods**: POST, GET, PATCH, DELETE

    ### Internal Endpoints (Write Operations)
    - **Path**: `/api/v1/internal/*`
    - **Access**: Network-restricted or authenticated
    - **Methods**: POST, PUT, PATCH, DELETE
    - **Security**: JWT Bearer tokens or API Keys required

    ## Authentication

    **JWT Bearer Tokens** (user accounts):
    - Register: POST `/api/v1/auth/register`
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /platforms/{platform}/environments/{environment}/versions:
    get:
      tags:
        - Config Versions
      summary: List all configuration versions
      description: Retrieve all configuration versions for an environment
      operationId: listVersions
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
      responses:
        '200':
          description: List of configuration versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Version'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
  /platforms/{platform}/environments/{environment}/versions/{version}:
    get:
      tags:
        - Config Versions
      summary: Get specific configuration version
      description: Retrieve a specific configuration version by timestamp
      operationId: getVersion
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/Version'
      responses:
        '200':
          description: Configuration version details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Version'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
...
  /platforms/{platform}/environments/{environment}/flags:
    get:
      tags:
        - Feature Flags
      summary: List all feature flags
      description: Retrieve all feature flags for an environment
      operationId: listFeatureFlags
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
      responses:
        '200':
          description: List of feature flags
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeatureFlag'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
  /platforms/{platform}/environments/{environment}/flags/{flagKey}:
    get:
      tags:
        - Feature Flags
      summary: Get specific feature flag
      description: Retrieve details of a specific feature flag
      operationId: getFeatureFlag
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
      responses:
        '200':
          description: Feature flag details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/FeatureFlag'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
  /platforms/{platform}/environments/{environment}/flags/{flagKey}/evaluate:
    get:
      tags:
        - Evaluation
      summary: Evaluate feature flag (public)
      description: |
        Evaluate a feature flag for a given context. This is a public endpoint for SDK usage.
        Returns the resolved value based on targeting rules.
      operationId: evaluateFlagPublic
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
        - name: userId
          in: query
          schema:
            type: string
          description: User identifier for targeting
        - name: country
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
          description: ISO 3166-1 alpha-2 country code
        - name: language
          in: query
          schema:
            type: string
            minLength: 2
            maxLength: 2
          description: ISO 639-1 language code
      responses:
        '200':
          description: Flag evaluation result
          headers:
            Cache-Control:
              schema:
                type: string
              description: CDN caching headers
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/EvaluationResult'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # FEATURE FLAGS (Internal Write)
  # ============================================================================
...
  /internal/platforms/{platform}/environments/{environment}/flags/{flagKey}/versions:
    get:
      tags:
        - Feature Flags
      summary: List flag versions
      description: Retrieve all versions of a specific feature flag
      operationId: listFlagVersions
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
      responses:
        '200':
          description: List of flag versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeatureFlag'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/platforms/{platform}/environments/{environment}/flags/{flagKey}/versions/{version}:
    get:
      tags:
        - Feature Flags
      summary: Get specific flag version
      description: Retrieve a specific version of a feature flag
      operationId: getFlagVersion
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Environment'
        - $ref: '#/components/parameters/FlagKey'
        - name: version
          in: path
          required: true
          schema:
            type: string
          description: Flag version number
      responses:
        '200':
          description: Flag version details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/FeatureFlag'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # FEATURE FLAG EVALUATION
  # ============================================================================
...
  /internal/cache/invalidate:
    post:
      tags:
        - Cache
      summary: Invalidate CloudFront cache
      description: Invalidate CDN cache with different granularity levels (global, platform, environment, or version).
      operationId: invalidateCache
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CacheInvalidation'
      responses:
        '200':
          description: Cache invalidation initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      invalidationId:
                        type: string
                      status:
                        type: string
                      paths:
                        type: array
                        items:
                          type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication
  parameters:
    Platform:
      name: platform
      in: path
      required: true
      schema:
        type: string
      description: Platform identifier
...
    Environment:
      name: environment
      in: path
      required: true
      schema:
        type: string
      description: Environment name
    Version:
      name: version
      in: path
      required: true
      schema:
        type: string
      description: Version label (human-readable identifier, e.g., "v1.0.0", "1.2.3"). This is the versionLabel field, not the auto-generated versionTimestamp.
    FlagKey:
      name: flagKey
      in: path
      required: true
      schema:
        type: string
      description: Feature flag key identifier
...
  schemas:
...
    Version:
      type: object
      required:
        - platform
        - environment
        - versionTimestamp
        - versionLabel
        - isStable
        - config
        - createdBy
        - createdAt
      properties:
        platform:
          type: string
        environment:
          type: string
        versionTimestamp:
          type: string
          format: date-time
          description: Auto-generated ISO-8601 timestamp for ordering
        versionLabel:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          description: Semantic version label (e.g., "1.0.0"). This is the primary identifier used in URLs.
        isStable:
          type: boolean
          default: false
          description: Production-ready flag
        config:
          $ref: '#/components/schemas/Config'
        createdBy:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    FeatureFlag:
      type: object
      description: |
        Feature Flag with 2-value model (valueA/valueB) and geographic targeting.
        For multi-variant testing (3+ variants), use Experiments instead.
      required:
        - platform
        - environment
        - flagKey
        - name
        - enabled
        - flagType
        - valueA
        - valueB
        - createdBy
        - createdAt
      properties:
        platform:
          type: string
        environment:
          type: string
        flagKey:
          type: string
          pattern: '^[a-z][a-z0-9-]*$'
          description: Unique identifier (lowercase, hyphenated, must start with letter)
        name:
          type: string
          description: Human-readable display name
        description:
          type: string
        enabled:
          type: boolean
          default: false
          description: Master switch - when false, always serves defaultValue
        flagType:
          type: string
          enum: [boolean, string, number]
          default: boolean
        valueA:
          oneOf:
            - type: boolean
            - type: string
            - type: number
          description: Primary value (served when conditions match)
        valueB:
          oneOf:
            - type: boolean
            - type: string
            - type: number
          description: Secondary value (fallback)
        targeting:
          $ref: '#/components/schemas/FlagTargeting'
        defaultValue:
          type: string
          enum: [A, B]
          default: B
          description: Value to serve when no targeting matches
        rolloutEnabled:
          type: boolean
          default: false
          description: Enable percentage-based rollout
        rolloutPercentageA:
          type: integer
          minimum: 0
          maximum: 100
          default: 100
          description: Percentage of users who receive valueA (0-100)
        rolloutPercentageB:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
          description: Percentage of users who receive valueB (auto-calculated as 100 - rolloutPercentageA)
        version:
          type: string
          description: Semantic version string
        isActive:
          type: boolean
          default: true
          description: Whether this version is the active one
        createdBy:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EvaluationResult:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Whether the flag is enabled for this context
        reason:
          type: string
          description: Human-readable explanation
        flagName:
          type: string
          description: Flag name (in batch evaluations)
    CacheInvalidation:
      type: object
      description: Cache invalidation request with different granularity levels.
      properties:
        platform:
          type: string
          description: Invalidate all caches for this platform
        environment:
          type: string
          description: Invalidate all caches for this environment (requires platform)
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          description: Semantic version label to invalidate (requires platform + environment)
        global:
          type: boolean
          description: Invalidate all caches globally
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...